Separate distances_simd_sse from distances_simd
