+++ knowhere/utils/FaissHookFvec.cpp
@@ -8,6 +8,7 @@
 #include "distances_simd.h"
 #include "distances_simd_avx.h"
 #include "distances_simd_avx512.h"
+#include "distances_simd_sse.h"
 #include "instruction_set.h"
 
 namespace faiss {
@@ -21,6 +22,11 @@ fvec_func_ptr fvec_inner_product = fvec_inner_product_avx;
 fvec_func_ptr fvec_L2sqr = fvec_L2sqr_avx;
 fvec_func_ptr fvec_L1 = fvec_L1_avx;
 fvec_func_ptr fvec_Linf = fvec_Linf_avx;
+fvec_norm_L2sqr_func_ptr fvec_norm_L2sqr = fvec_norm_L2sqr_sse;
+fvec_L2sqr_ny_func_ptr fvec_L2sqr_ny = fvec_L2sqr_ny_sse;
+fvec_inner_products_ny_func_ptr fvec_inner_products_ny = fvec_inner_products_ny_sse;
+fvec_madd_func_ptr fvec_madd = fvec_madd_sse;
+fvec_madd_and_argmin_func_ptr fvec_madd_and_argmin = fvec_madd_and_argmin_sse;
 
 /*****************************************************************************/
 
@@ -53,6 +59,12 @@ void hook_fvec(std::string& simd_type) {
         fvec_L1 = fvec_L1_avx512;
         fvec_Linf = fvec_Linf_avx512;
 
+        fvec_norm_L2sqr = fvec_norm_L2sqr_sse;
+        fvec_L2sqr_ny = fvec_L2sqr_ny_sse;
+        fvec_inner_products_ny = fvec_inner_products_ny_sse;
+        fvec_madd = fvec_madd_sse;
+        fvec_madd_and_argmin = fvec_madd_and_argmin_sse;
+
         simd_type = "AVX512";
     } else if (faiss_use_avx2 && cpu_support_avx2()) {
         /* for IVFFLAT */
@@ -60,6 +72,12 @@ void hook_fvec(std::string& simd_type) {
         fvec_L2sqr = fvec_L2sqr_avx;
         fvec_L1 = fvec_L1_avx;
         fvec_Linf = fvec_Linf_avx;
+
+        fvec_norm_L2sqr = fvec_norm_L2sqr_sse;
+        fvec_L2sqr_ny = fvec_L2sqr_ny_sse;
+        fvec_inner_products_ny = fvec_inner_products_ny_sse;
+        fvec_madd = fvec_madd_sse;
+        fvec_madd_and_argmin = fvec_madd_and_argmin_sse;
 
         simd_type = "AVX2";
     } else if (faiss_use_sse4_2 && cpu_support_sse4_2()) {
@@ -69,6 +87,12 @@ void hook_fvec(std::string& simd_type) {
         fvec_L1 = fvec_L1_sse;
         fvec_Linf = fvec_Linf_sse;
 
+        fvec_norm_L2sqr = fvec_norm_L2sqr_sse;
+        fvec_L2sqr_ny = fvec_L2sqr_ny_sse;
+        fvec_inner_products_ny = fvec_inner_products_ny_sse;
+        fvec_madd = fvec_madd_sse;
+        fvec_madd_and_argmin = fvec_madd_and_argmin_sse;
+
         simd_type = "SSE4_2";
     } else {
         /* for IVFFLAT */
@@ -76,6 +100,12 @@ void hook_fvec(std::string& simd_type) {
         fvec_L2sqr = fvec_L2sqr_ref;
         fvec_L1 = fvec_L1_ref;
         fvec_Linf = fvec_Linf_ref;
+
+        fvec_norm_L2sqr = fvec_norm_L2sqr_ref;
+        fvec_L2sqr_ny = fvec_L2sqr_ny_ref;
+        fvec_inner_products_ny = fvec_inner_products_ny_ref;
+        fvec_madd = fvec_madd_ref;
+        fvec_madd_and_argmin = fvec_madd_and_argmin_ref;
 
         simd_type = "REF";
     }
